객실 수정하기 수정중인데 이거 심상치않은데?

+++ src/main/resources/static/js/roomUpdate.js
@@ -1,21 +1,35 @@
-  $(function() {
-	$(document).on("click", ".removeImg", (event)=> {
-		$(event.target).parent().remove();
-		const room_img_sid = $(event.target).val();
-		const current_img = $(event.target).closest(".currentImg").val();
-		const ajaxObj = {
-			url : API_LIST.DELETE_ROOM_IMG,
-			method : "delete",
-			param : {
-				room_img_sid : room_img_sid,
-				current_img : current_img
-			},
-			successFn : (resultResponse) => {
-				if(resultResponse.success){
-					swalCall("성공",resultResponse.message,"success");
-				}
-			}
-		}
-		ajaxCall(ajaxObj);
-	});
+$(function () {
+    $(document).on("click", ".removeImg", (event) => {
+        const room_img_sid = $(event.target).val();
+        const current_img_box = $(event.target).closest(".current_img_box");
+        const current_img = current_img_box.find(".currentImg").val();
+        const thenFn = (result) => {
+            if (result.isConfirmed) {
+                const ajaxObj = {
+                    url: API_LIST.DELETE_ROOM_IMG,
+                    method: "delete",
+                    param: {
+                        room_img_sid: room_img_sid,
+                        current_img: current_img
+                    },
+                    successFn: () => {
+                        const thenFn = () =>{
+                            location.reload();
+                        }
+                        swalCall("성공", "이미지 삭제 완료", "success",thenFn);
+                    }
+                }
+                ajaxCall(ajaxObj);
+            } else {
+                return;
+            }
+        }
+        swalCall("이미지 삭제", "해당 이미지를 삭제하시겠습니까?", "question", thenFn, "예", true);
+    });
+
+    $(".room_update_btn").click(() => {
+        const roomUpdateForm = $(".roomUpdateForm");
+
+        roomUpdateForm.submit();
+    });
 });
